requirements.txt
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
findspark
pyspark-stubs
geopandas
shapely
gtfs-realtime-bindings